Moved autoreplies to autoreply.py
